<fix>:修复了qspi无法写入的问题,原因在于qspi_send_cmd函数的地址参数应该是uint32_t类型的,而不是uint16_t类型的.

+++ drv/inc/drv_qspi.h
@@ -37,32 +37,32 @@
 #define QSPI_BK1_CLK_GPIO_PORT          GPIOB
 #define QSPI_BK1_CLK_GPIO_PIN           GPIO_PIN_2
 #define QSPI_BK1_CLK_GPIO_AF            GPIO_AF9_QUADSPI
-#define QSPI_BK1_CLK_GPIO_CLK_ENABLE()  do{ __HAL_RCC_GPIOB_CLK_ENABLE; }while(0)   /* PB口时钟使能 */
+#define QSPI_BK1_CLK_GPIO_CLK_ENABLE()  do{ __HAL_RCC_GPIOB_CLK_ENABLE; }while(0)   /* PB锟斤拷时锟斤拷使锟斤拷 */
 
 #define QSPI_BK1_NCS_GPIO_PORT          GPIOB
 #define QSPI_BK1_NCS_GPIO_PIN           GPIO_PIN_6
 #define QSPI_BK1_NCS_GPIO_AF            GPIO_AF10_QUADSPI
-#define QSPI_BK1_NCS_GPIO_CLK_ENABLE()  do{ __HAL_RCC_GPIOB_CLK_ENABLE; }while(0)   /* PB口时钟使能 */
+#define QSPI_BK1_NCS_GPIO_CLK_ENABLE()  do{ __HAL_RCC_GPIOB_CLK_ENABLE; }while(0)   /* PB锟斤拷时锟斤拷使锟斤拷 */
 
 #define QSPI_BK1_IO0_GPIO_PORT          GPIOD
 #define QSPI_BK1_IO0_GPIO_PIN           GPIO_PIN_11
 #define QSPI_BK1_IO0_GPIO_AF            GPIO_AF9_QUADSPI
-#define QSPI_BK1_IO0_GPIO_CLK_ENABLE()  do{ __HAL_RCC_GPIOD_CLK_ENABLE; }while(0)   /* PD口时钟使能 */
+#define QSPI_BK1_IO0_GPIO_CLK_ENABLE()  do{ __HAL_RCC_GPIOD_CLK_ENABLE; }while(0)   /* PD锟斤拷时锟斤拷使锟斤拷 */
 
 #define QSPI_BK1_IO1_GPIO_PORT          GPIOD
 #define QSPI_BK1_IO1_GPIO_PIN           GPIO_PIN_12
 #define QSPI_BK1_IO1_GPIO_AF            GPIO_AF9_QUADSPI
-#define QSPI_BK1_IO1_GPIO_CLK_ENABLE()  do{ __HAL_RCC_GPIOD_CLK_ENABLE; }while(0)   /* PD口时钟使能 */
+#define QSPI_BK1_IO1_GPIO_CLK_ENABLE()  do{ __HAL_RCC_GPIOD_CLK_ENABLE; }while(0)   /* PD锟斤拷时锟斤拷使锟斤拷 */
 #if 0
 #define QSPI_BK1_IO2_GPIO_PORT          GPIOD
 #define QSPI_BK1_IO2_GPIO_PIN           GPIO_PIN_13
 #define QSPI_BK1_IO2_GPIO_AF            GPIO_AF9_QUADSPI
-#define QSPI_BK1_IO2_GPIO_CLK_ENABLE()  do{ __HAL_RCC_GPIOD_CLK_ENABLE; }while(0)   /* PD口时钟使能 */
+#define QSPI_BK1_IO2_GPIO_CLK_ENABLE()  do{ __HAL_RCC_GPIOD_CLK_ENABLE; }while(0)   /* PD锟斤拷时锟斤拷使锟斤拷 */
 
 #define QSPI_BK1_IO3_GPIO_PORT          GPIOE
 #define QSPI_BK1_IO3_GPIO_PIN           GPIO_PIN_2
 #define QSPI_BK1_IO3_GPIO_AF            GPIO_AF9_QUADSPI
-#define QSPI_BK1_IO3_GPIO_CLK_ENABLE()  do{ __HAL_RCC_GPIOE_CLK_ENABLE; }while(0)   /* PE口时钟使能 */
+#define QSPI_BK1_IO3_GPIO_CLK_ENABLE()  do{ __HAL_RCC_GPIOE_CLK_ENABLE; }while(0)   /* PE锟斤拷时锟斤拷使锟斤拷 */
 #else
 #define QSPI_BK1_IO2_GPIO_PORT   GPIOE
 #define QSPI_BK1_IO2_GPIO_PIN    GPIO_PIN_2
@@ -78,7 +78,7 @@
 
 uint8_t qspi_wait_flag(uint32_t flag, uint8_t sta, uint32_t wtime);
 uint8_t qspi_init(void);
-void qspi_send_cmd(uint8_t cmd, uint16_t addr, uint8_t mode, uint8_t dmcycle);
+void qspi_send_cmd(uint8_t cmd, uint32_t addr, uint8_t mode, uint8_t dmcycle);
 uint8_t qspi_receive(uint8_t* buf, uint32_t datalen);
 uint8_t qspi_transmit(uint8_t* buf, uint32_t datalen);
